Move AppRoutesRedirect
