+++ src/client/config/routes.ts
@@ -1,0 +1,50 @@
+import { generatePath } from 'react-router-dom';
+import { AppRoutes } from '../../universal/config/routes';
+
+export const AppRoutesRedirect = [
+  {
+    from: '/burgerzaken/document/:id',
+    to: AppRoutes['BURGERZAKEN/ID-KAART'],
+  },
+  {
+    from: '/inkomen-en-stadspas/stadspas/aanvraag/:id',
+    to: AppRoutes['STADSPAS/AANVRAAG'],
+  },
+  {
+    from: '/inkomen-en-stadspas/stadspas/saldo/:id',
+    to: AppRoutes['STADSPAS/SALDO'],
+  },
+  {
+    from: '/inkomen-en-stadspas/bijstandsuitkering/:id',
+    to: AppRoutes['INKOMEN/BIJSTANDSUITKERING'],
+  },
+  {
+    from: '/inkomen-en-stadspas/uitkeringsspecificaties/jaaropgaven',
+    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
+      variant: 'jaaropgave',
+    }),
+  },
+  {
+    from: '/inkomen/uitkeringsspecificaties/jaaropgaven',
+    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
+      variant: 'jaaropgave',
+    }),
+  },
+  {
+    from: '/inkomen-en-stadspas/uitkeringsspecificaties/',
+    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
+      variant: 'uitkering',
+    }),
+  },
+  {
+    from: '/inkomen/uitkeringsspecificaties/',
+    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
+      variant: 'uitkering',
+    }),
+  },
+  {
+    from: '/inkomen-en-stadspas/tozo/:version/:id',
+    to: AppRoutes['INKOMEN/TOZO'],
+  },
+  { from: '/inkomen-en-stadspas', to: AppRoutes.INKOMEN },
+];
