+++ src/universal/config/routes.ts
@@ -71,54 +71,6 @@ export const AppRoutes: Record<string, string> = {
   ZAAK_STATUS: '/zaak-status',
 } as const;
 
-export const AppRoutesRedirect = [
-  {
-    from: '/burgerzaken/document/:id',
-    to: AppRoutes['BURGERZAKEN/ID-KAART'],
-  },
-  {
-    from: '/inkomen-en-stadspas/stadspas/aanvraag/:id',
-    to: AppRoutes['STADSPAS/AANVRAAG'],
-  },
-  {
-    from: '/inkomen-en-stadspas/stadspas/saldo/:id',
-    to: AppRoutes['STADSPAS/SALDO'],
-  },
-  {
-    from: '/inkomen-en-stadspas/bijstandsuitkering/:id',
-    to: AppRoutes['INKOMEN/BIJSTANDSUITKERING'],
-  },
-  {
-    from: '/inkomen-en-stadspas/uitkeringsspecificaties/jaaropgaven',
-    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
-      variant: 'jaaropgave',
-    }),
-  },
-  {
-    from: '/inkomen/uitkeringsspecificaties/jaaropgaven',
-    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
-      variant: 'jaaropgave',
-    }),
-  },
-  {
-    from: '/inkomen-en-stadspas/uitkeringsspecificaties/',
-    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
-      variant: 'uitkering',
-    }),
-  },
-  {
-    from: '/inkomen/uitkeringsspecificaties/',
-    to: generatePath(AppRoutes['INKOMEN/SPECIFICATIES'], {
-      variant: 'uitkering',
-    }),
-  },
-  {
-    from: '/inkomen-en-stadspas/tozo/:version/:id',
-    to: AppRoutes['INKOMEN/TOZO'],
-  },
-  { from: '/inkomen-en-stadspas', to: AppRoutes.INKOMEN },
-];
-
 export const PublicRoutes: string[] = [
   AppRoutes.API_LOGIN,
   AppRoutes.API1_LOGIN,
